Move segment managers inline and improve naming
